+++ ecs/archetype.go
@@ -60,6 +60,14 @@ func (a *archetype) Get(index int, id ID) unsafe.Pointer {
 	return a.components[a.indices[id]].Get(uint32(index))
 }
 
+// GetUnsafe returns the component with the given ID at the given index,
+// without checking if the entity contains that component.
+//
+// This is used by queries, where the entity is guaranteed to be in the archetype.
+func (a *archetype) GetUnsafe(index int, id ID) unsafe.Pointer {
+	return a.components[a.indices[id]].Get(uint32(index))
+}
+
 // Add adds an entity with components to the archetype
 func (a *archetype) Add(entity Entity, components ...Component) uint32 {
 	if len(components) != len(a.ids) {